@@ -3,91 +3,111 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@babel/core": "^7.7.2",
-    "@babel/plugin-proposal-class-properties": "^7.7.0",
-    "@babel/plugin-proposal-export-default-from": "^7.5.2",
-    "@babel/plugin-proposal-export-namespace-from": "^7.5.2",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-transform-destructuring": "^7.6.0",
-    "@babel/plugin-transform-runtime": "^7.6.2",
-    "@babel/preset-env": "^7.7.1",
-    "@babel/preset-react": "^7.7.0",
+    "@babel/core": "^7.8.6",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-export-default-from": "^7.8.3",
+    "@babel/plugin-proposal-export-namespace-from": "^7.8.3",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-destructuring": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.8.3",
+    "@babel/preset-env": "^7.8.6",
+    "@babel/preset-react": "^7.8.3",
+    "@loadable/component": "^5.12.0",
     "babel-loader": "^8.0.6",
-    "babel-plugin-react-intl": "^5.1.4",
-    "browser-env": "^3.2.5",
-    "case-sensitive-paths-webpack-plugin": "^2.2.0",
-    "classnames": "^2.2.5",
-    "connected-react-router": "^6.6.1",
-    "core-js": "^3.4.3",
+    "babel-plugin-react-intl": "^5.1.18",
+    "browser-env": "^3.3.0",
+    "case-sensitive-paths-webpack-plugin": "^2.3.0",
+    "classnames": "^2.2.6",
+    "connected-react-router": "^6.7.0",
+    "core-js": "^3.6.4",
     "country-telephone-data": "^0.6.0",
     "create-file-webpack": "^1.0.2",
-    "cross-env": "^6.0.3",
-    "css-loader": "^3.2.0",
-    "deep-equal": "^1.0.1",
+    "cross-env": "^7.0.1",
+    "css-loader": "^3.4.2",
+    "deep-equal": "^2.0.1",
     "detect-browser": "^4.8.0",
     "dotenv": "^8.2.0",
     "dotenv-expand": "^5.1.0",
-    "draft-js": "^0.11.2",
-    "draftjs-to-html": "^0.8.4",
+    "draft-js": "^0.11.4",
+    "draftjs-to-html": "^0.9.1",
     "druid-query": "git+https://github.com/DeyLak/node-druid-query.git",
-    "file-loader": "^5.0.2",
+    "immutable": "4.0.0-rc.12",
+    "file-loader": "^5.1.0",
     "fs-extra": "^8.1.0",
     "gulp": "^4.0.2",
-    "gulp-rename": "^1.2.2",
+    "gulp-rename": "^2.0.0",
     "gulp-template": "^5.0.0",
-    "history": "^4.7.2",
-    "html-to-react": "^1.3.4",
+    "history": "^4.10.1",
+    "html-to-react": "^1.4.2",
     "html-webpack-plugin": "4.0.0-beta.11",
     "lodash": "^4.17.15",
-    "memoize-one": "^5.0.3",
-    "mini-css-extract-plugin": "^0.8.0",
-    "moment": "^2.22.0",
+    "memoize-one": "^5.1.1",
+    "mini-css-extract-plugin": "^0.9.0",
+    "moment": "^2.24.0",
     "moment-locales-webpack-plugin": "^1.1.2",
     "null-loader": "^3.0.0",
-    "object-assign": "4.1.1",
-    "object-hash": "^2.0.0",
+    "object-assign": "^4.1.1",
+    "object-hash": "^2.0.3",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
-    "postcss-flexbugs-fixes": "^4.1.0",
+    "postcss-calc": "^7.0.2",
+    "postcss-flexbugs-fixes": "^4.2.0",
     "postcss-global-import": "^1.0.6",
     "postcss-import": "^12.0.1",
     "postcss-loader": "^3.0.0",
     "postcss-preset-env": "^6.7.0",
-    "postcss-safe-parser": "^4.0.1",
-    "promise": "^8.0.3",
-    "qhistory": "^1.0.2",
-    "qs": "^6.7.0",
+    "postcss-safe-parser": "^4.0.2",
+    "promise": "^8.1.0",
+    "qhistory": "^1.0.3",
+    "qs": "^6.9.1",
     "raf": "^3.4.1",
-    "react": "^16.8.1",
-    "react-dev-utils": "^9.1.0",
-    "react-dnd": "^2.6.0",
-    "react-dnd-html5-backend": "^2.6.0",
-    "react-dom": "^16.8.1",
+    "react": "^16.13.0",
+    "react-dev-utils": "9.1.0",
+    "react-dnd": "^10.0.2",
+    "react-dnd-html5-backend": "^10.0.2",
+    "react-dom": "^16.13.0",
     "react-intl": "3.0.0-beta-5",
     "react-intl-po": "^2.2.2",
-    "react-linkify": "^0.2.2",
-    "react-loadable": "^5.3.1",
-    "react-modal": "^3.3.2",
-    "react-moment-proptypes": "^1.5.0",
-    "react-redux": "^7.1.3",
+    "react-linkify": "0.2.2",
+    "react-modal": "^3.11.2",
+    "react-moment-proptypes": "^1.7.0",
+    "react-redux": "^7.2.0",
     "react-router-dom": "^5.1.2",
     "react-transition-group": "^4.3.0",
-    "recharts": "1.7.1",
+    "recharts": "2.0.0-beta.1",
     "reduce-reducers": "^1.0.4",
-    "redux": "^4.0.4",
+    "redux": "^4.0.5",
     "redux-batched-actions": "^0.4.1",
-    "redux-restify": "0.1.5",
-    "redux-thunk": "^2.2.0",
+    "redux-restify": "^0.6.6",
+    "redux-thunk": "^2.3.0",
     "regenerator-runtime": "^0.13.3",
-    "sanitize-html": "^1.19.0",
-    "scroll-into-view-if-needed": "^2.1.5",
-    "style-loader": "1.0.2",
-    "sw-precache-webpack-plugin": "^0.11.5",
-    "terser-webpack-plugin": "^2.2.1",
+    "sanitize-html": "^1.22.0",
+    "scroll-into-view-if-needed": "^2.2.24",
+    "seamless-immutable": "^7.1.4",
+    "style-loader": "^1.1.3",
+    "sw-precache-webpack-plugin": "^1.0.0",
+    "terser-webpack-plugin": "^2.3.5",
+    "typescript": "^3.8.3",
     "url-loader": "^3.0.0",
-    "webpack": "4.41.2",
-    "webpack-dev-server": "^3.9.0",
+    "webpack": "^4.42.0",
+    "webpack-dev-server": "^3.10.3",
     "webpack-manifest-plugin": "^2.2.0",
     "whatwg-fetch": "^3.0.0"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^2.22.0",
+    "@typescript-eslint/parser": "^2.22.0",
+    "babel-eslint": "^10.0.3",
+    "eslint": "^6.7.1",
+    "eslint-config-react-app": "^5.0.2",
+    "eslint-plugin-flowtype": "^4.5.2",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.17.0",
+    "eslint-plugin-react-hooks": "^2.3.0",
+    "pre-commit": "^1.2.2",
+    "react-styleguidist": "^10.6.2",
+    "stylelint": "^13.2.0",
+    "stylelint-config-standard": "^20.0.0"
   },
   "scripts": {
     "build-gulp": "node scripts/build-gulp.js",
