@@ -12,31 +12,48 @@ import { intlObject } from '$trood/localeService'
 
 class EnchancedSwitch extends PureComponent {
   static propTypes = {
+    /** class name for styling component */
     className: PropTypes.string,
+    /** class name for styling label */
     labelClassName: PropTypes.string,
+    /** class name for styling disabled label */
     disabledLabelClassName: PropTypes.string,
-
+    /** type is one of  INNER_INPUT_TYPES.checkbox, INNER_INPUT_TYPES.radio */
     type: PropTypes.oneOf(Object.values(INNER_INPUT_TYPES)),
+    /** disabled or not */
     disabled: PropTypes.bool,
+    /** stop propagation or not */
     stopPropagation: PropTypes.bool,
-
+    /** switched or not */
     switched: PropTypes.bool.isRequired,
+    /** view switched component */
     switchedComponent: PropTypes.node.isRequired,
+    /** view un switched component */
     unSwitchedComponent: PropTypes.node,
+    /** label text */
     label: PropTypes.node,
+    /** label position is one of LABEL_POSITION_TYPES.right, LABEL_POSITION_TYPES.left */
     labelPosition: PropTypes.oneOf(Object.values(LABEL_POSITION_TYPES)),
+    /** second label text */
     secondLabel: PropTypes.node,
+    /** second label position is one of LABEL_POSITION_TYPES.right, LABEL_POSITION_TYPES.left */
     secondLabelPosition: PropTypes.oneOf(Object.values(LABEL_POSITION_TYPES)),
-
+    /** errors text */
     errors: PropTypes.arrayOf(PropTypes.node),
+    /** show text errors or not */
     showTextErrors: PropTypes.bool,
+    /** validate settings */
     validate: PropTypes.shape({
+      /** check on blur or not */
       checkOnBlur: PropTypes.bool,
+      /** required or not */
       required: PropTypes.bool,
     }),
-
+    /** onChange function */
     onChange: PropTypes.func,
+    /** onValid function */
     onValid: PropTypes.func,
+    /** onInvalid function */
     onInvalid: PropTypes.func,
   }
 
@@ -74,9 +91,9 @@ class EnchancedSwitch extends PureComponent {
     this.validate(this.props.switched)
   }
 
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.switched !== this.props.switched) {
-      this.validate(nextProps.switched)
+  componentDidUpdate(prevProps) {
+    if (prevProps.switched !== this.props.switched) {
+      this.validate(this.props.switched)
     }
   }
 
@@ -127,14 +144,13 @@ class EnchancedSwitch extends PureComponent {
         tabIndex: 0,
         className: classNames(style.root, className, disabled && style.disabled),
         'data-cy': label,
-        onClick: !disabled && ((e) => {
-          if (stopPropagation) {
-            e.stopPropagation()
-          }
-          onChange(!switched)
-        }),
-        onBlur: !disabled && (() => this.setState({ wasBlured: true })),
-        onFocus: !disabled && (() => this.setState({ wasBlured: false })),
+        onClick: !disabled ?
+          e => {
+            if (stopPropagation) e.stopPropagation()
+            onChange(!switched)
+          } : undefined,
+        onBlur: !disabled ? () => this.setState({ wasBlured: true }) : undefined,
+        onFocus: !disabled ? () => this.setState({ wasBlured: false }) : undefined,
       }} >
         <input {...{
           type,
@@ -184,5 +200,7 @@ class EnchancedSwitch extends PureComponent {
     )
   }
 }
+
+export { INNER_INPUT_TYPES, LABEL_POSITION_TYPES } from './constants'
 
 export default EnchancedSwitch
