@@ -1,16 +1,22 @@
-import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
+import PropTypes from 'prop-types'
 import classNames from 'classnames'
 
 import style from './index.css'
 
 import LoadingIndicator from '$trood/components/LoadingIndicator'
 
+/**
+ * Component for loading block container.
+ */
 
 class LoadingBlockContainer extends PureComponent {
   static propTypes = {
+    /** class name for styling component */
     className: PropTypes.string,
+    /** show loading indicator or not */
     isBlocked: PropTypes.bool,
+    /** children node */
     children: PropTypes.node,
   }
 
