@@ -1,21 +1,26 @@
-import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
-import classNames from 'classnames'
+import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
+import classNames from 'classnames'
 
 import style from './index.css'
+
+import TIcon, { ICONS_TYPES } from '$trood/components/TIcon'
+
 import {
   BUTTON_TYPES,
   BUTTON_COLORS,
   BUTTON_SPECIAL_TYPES,
 } from './constants'
 
-import TIcon from '$trood/components/TIcon'
-import { ICONS_TYPES } from '$trood/components/TIcon/constants'
-
-
 const BUTTON_SPECIAL_ICONS = {
   [BUTTON_SPECIAL_TYPES.add]: (
+    <TIcon {...{
+      type: ICONS_TYPES.plus,
+      className: style.addIcon,
+    }} />
+  ),
+  [BUTTON_SPECIAL_TYPES.addFill]: (
     <TIcon {...{
       type: ICONS_TYPES.plus,
       className: style.addIcon,
@@ -62,20 +67,56 @@ const BUTTON_SPECIAL_ICONS = {
       className: style.textIcon,
     }} />
   ),
+  [BUTTON_SPECIAL_TYPES.edit]: (
+    <TIcon {...{
+      type: ICONS_TYPES.edit,
+      className: style.textIcon,
+    }} />
+  ),
+  [BUTTON_SPECIAL_TYPES.download]: (
+    <TIcon {...{
+      type: ICONS_TYPES.download,
+      className: style.textIcon,
+    }} />
+  ),
+  [BUTTON_SPECIAL_TYPES.upload]: (
+    <TIcon {...{
+      type: ICONS_TYPES.upload,
+      className: style.textIcon,
+    }} />
+  ),
 }
+
+/**
+ * Component for output Button.
+ */
 
 class TButton extends PureComponent {
   static propTypes = {
+    /** type is one of BUTTON_TYPES.text, BUTTON_TYPES.border, BUTTON_TYPES.fill */
     type: PropTypes.oneOf(Object.values(BUTTON_TYPES)),
+    /** special type is one of BUTTON_TYPES.add, BUTTON_TYPES.addFill, BUTTON_TYPES.minus, BUTTON_TYPES.delete,
+     * BUTTON_TYPES.text, BUTTON_TYPES.icon, BUTTON_TYPES.arrowLeft, BUTTON_TYPES.arrowRight, BUTTON_TYPES.attach,
+     * BUTTON_TYPES.edit, BUTTON_TYPES.download, BUTTON_TYPES.upload */
     specialType: PropTypes.oneOf(Object.values(BUTTON_SPECIAL_TYPES)),
+    /** color is one of BUTTON_COLORS.red, BUTTON_COLORS.blue, BUTTON_COLORS.white, BUTTON_COLORS.gray,
+     * BUTTON_COLORS.orange, BUTTON_COLORS.green */
     color: PropTypes.oneOf(Object.values(BUTTON_COLORS)),
+    /** tab index number*/
     tabIndex: PropTypes.number,
+    /** link */
     link: PropTypes.string,
+    /** label text */
     label: PropTypes.node,
+    /** set tooltip */
     tooltip: PropTypes.string,
+    /** thin or not */
     thin: PropTypes.bool,
+    /** disabled or not */
     disabled: PropTypes.bool,
+    /** onClick function */
     onClick: PropTypes.func,
+    /** class name for styling component */
     className: PropTypes.string,
   }
 
