@@ -12,20 +12,33 @@ import {
 
 import style from './index.css'
 
+/**
+ * Component for output icon.
+ */
 
 class TIcon extends PureComponent {
   static propTypes = {
+    /** class name for styling component */
+    className: PropTypes.string,
+    /** width and height size icon in px */
     size: PropTypes.number,
+    /** all types you can see below in example */
     type: PropTypes.oneOf(Object.keys(ICONS_TYPES)),
+    /** default type for rotate you can see in constants, or send number in deg */
     rotate: PropTypes.oneOfType([
       PropTypes.oneOf(Object.values(ROTATE_TYPES)),
       PropTypes.number,
     ]),
+    /** label icon */
     label: PropTypes.node,
+    /** label position is one of LABEL_POSITION_TYPES.up, LABEL_POSITION_TYPES.right, LABEL_POSITION_TYPES.down,
+     * LABEL_POSITION_TYPES.left, LABEL_POSITION_TYPES.tooltip */
     labelPosition: PropTypes.oneOf(Object.keys(LABEL_POSITION_TYPES)),
-    className: PropTypes.string,
+    /** color icon */
     color: PropTypes.string,
+    /** disabled or not */
     disabled: PropTypes.bool,
+    /** onClick function */
     onClick: PropTypes.func,
   }
 
@@ -55,9 +68,15 @@ class TIcon extends PureComponent {
       width: size,
       height: size,
     }
+
     return (
       <div {...{
-        className: classNames(style.root, className),
+        className: classNames(
+          style.root,
+          className,
+          labelPosition === LABEL_POSITION_TYPES.down || labelPosition === LABEL_POSITION_TYPES.up ?
+            style['root-column'] : '',
+        ),
         'data-cy': type,
         onClick: disabled ? undefined : onClick,
         style: {
