@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types'
-import React from 'react'
+import React, { PureComponent} from 'react'
 import { Route } from 'react-router-dom'
 import classNames from 'classnames'
 
@@ -10,137 +10,119 @@ import urlSchema, {
   HEADER_HIDE,
 } from '../../urlSchema'
 
-import TIcon from '$trood/components/TIcon'
-
 import { getNestedObjectField } from '$trood/helpers/nestedObjects'
-import localeService, { intlObject } from '$trood/localeService'
 
-import PageNavLink from '../PageNavLink'
 import { HEADER_TYPES, MIN_MENU_RENDERERS } from './constants'
 
+import MainLink from './components/MainLink'
+import AdditionalLink from './components/AdditionalLink'
 
-const HeaderMenu = ({
-  className = '',
-  type = HEADER_TYPES.left,
 
-  basePath = [],
-  menuRenderers = {},
+class HeaderMenu extends PureComponent {
+  static propTypes = {
+    type: PropTypes.oneOf(Object.values(HEADER_TYPES)),
+    basePath: PropTypes.array,
+    menuRenderers: PropTypes.object,
+    additionalLinks: PropTypes.array,
+    autoResize: PropTypes.bool,
+  }
 
-  additionalLinks = [],
-  LinkComponent,
-  iconClassName,
-  linkClassName,
-  linkActiveClassName,
+  static defaultProps = {
+    type: HEADER_TYPES.left,
+    basePath: [],
+    menuRenderers: {},
+    additionalLinks: [],
+    autoResize: true,
+  }
 
-  autoResize = true,
-}) => {
-  if (Object.keys(menuRenderers).length < MIN_MENU_RENDERERS && !additionalLinks.length) return null
-  const preAdditionalLinks = additionalLinks.filter(link => link && link.pre)
-  const postAdditionalLinks = additionalLinks.filter(link => link && !link.pre)
-  const mapAdditionalLinks = (link, key) => {
+  constructor(props) {
+    super(props)
+
+    this.state = {
+      open: false,
+    }
+  }
+
+  render() {
+    const {
+      className = '',
+      type = HEADER_TYPES.left,
+
+      basePath = [],
+      menuRenderers = {},
+
+      additionalLinks = [],
+
+      autoResize = true,
+    } = this.props
+
+    const { open } = this.state
+
+    if (Object.keys(menuRenderers).length < MIN_MENU_RENDERERS && !additionalLinks.length) return null
+    const preAdditionalLinks = additionalLinks.filter(link => link && link.pre)
+    const postAdditionalLinks = additionalLinks.filter(link => link && !link.pre)
+
     return (
-      <div {...{
-        key,
-        onClick: link.onClick && (() => link.onClick()),
-        className: classNames(
-          linkClassName,
-          style.url,
-          link.active && style.active,
-          link.active && linkActiveClassName,
-          link.className,
-          link.active && link.activeClassName,
-        ),
-      }}>
-        {
-          !LinkComponent &&
-          <React.Fragment>
-            {link.iconType &&
-              <TIcon {...{
-                className: classNames(style.icon, iconClassName),
-                type: link.iconType,
+      <Route render={match => {
+        const currentSchemaConfig = getNestedObjectField(
+          urlSchemaConfig,
+          [].concat(basePath)
+            .map(name => [name, 'pages'])
+            .reduce((a, b) => a.concat(b), []),
+        )
+        const currentBaseUrl = getNestedObjectField(urlSchema, basePath)
+
+        const routeToRenderedKeysReduceFunc = (schemaConfig, prevKey) => (memo, key) => {
+          const currentConfig = schemaConfig[key]
+          if (currentConfig.header === HEADER_HIDE) return memo
+          if (currentConfig.header === HEADER_SHOW) return [...memo, prevKey ? [].concat(prevKey, key) : key]
+          return memo
+        }
+        const headerKeysToRender = Object.keys(currentSchemaConfig).reduce(
+          routeToRenderedKeysReduceFunc(currentSchemaConfig),
+          [],
+        )
+
+        return (
+          <div {...{
+            className: classNames(
+              style.root,
+              style[type],
+              className,
+              autoResize && style.autoResize,
+              open && style.open,
+            ),
+            onClick: () => this.setState({ open: !open }),
+          }}>
+            {preAdditionalLinks.map((link, key) => (
+              <AdditionalLink {...{
+                ...this.props,
+                link,
+                key,
               }} />
-            }
-            <span>{link.label}</span>
-          </React.Fragment>
-        }
-        {
-          LinkComponent &&
-          <LinkComponent {...{
-            icon: link.iconType,
-            label: link.label,
-          }} />
-        }
-      </div>
+            ))}
+            {headerKeysToRender.map(key => {
+              return (
+                <MainLink {...{
+                  ...this.props,
+                  key,
+                  navKey: key,
+                  currentBaseUrl,
+                }} />
+              )
+            })}
+            {postAdditionalLinks.map((link, key) => (
+              <AdditionalLink {...{
+                ...this.props,
+                link,
+                key,
+              }} />
+            ))}
+          </div>
+        )
+      }} />
     )
   }
-  return (
-    <Route render={() => {
-      const currentSchemaConfig = getNestedObjectField(
-        urlSchemaConfig,
-        [].concat(basePath)
-          .map(name => [name, 'pages'])
-          .reduce((a, b) => a.concat(b), []),
-      )
-      const currentBaseUrl = getNestedObjectField(urlSchema, basePath)
-
-      const routeToRenderedKeysReduceFunc = (schemaConfig, prevKey) => (memo, key) => {
-        const currentConfig = schemaConfig[key]
-        if (currentConfig.header === HEADER_HIDE) return memo
-        if (currentConfig.header === HEADER_SHOW) return [...memo, prevKey ? [].concat(prevKey, key) : key]
-        return memo
-      }
-      const headerKeysToRender = Object.keys(currentSchemaConfig).reduce(
-        routeToRenderedKeysReduceFunc(currentSchemaConfig),
-        [],
-      )
-      return (
-        <div className={classNames(style.root, style[type], className, autoResize && style.autoResize)}>
-          {preAdditionalLinks.map(mapAdditionalLinks)}
-          {headerKeysToRender.map(key => {
-            const currentRenderer = getNestedObjectField(menuRenderers, key) || menuRenderers[key[key.length - 1]]
-            if (!currentRenderer) return null
-            let currentLabel = currentRenderer.label
-            const currentMessage = localeService.localeMessages[currentRenderer.localeMessageId]
-            if (currentRenderer.localeMessageId && intlObject.intl && currentMessage) {
-              currentLabel = intlObject.intl.formatMessage(currentMessage)
-            }
-            const currentIconType = currentRenderer.iconType
-            return (
-              <PageNavLink {...{
-                key,
-                navKey: key,
-                baseUrl: currentBaseUrl,
-                activeClassName: classNames(style.active, linkActiveClassName),
-                className: classNames(style.url, linkClassName),
-              }}>
-                {
-                  !LinkComponent &&
-                  <React.Fragment>
-                    {currentIconType &&
-                      <TIcon className={classNames(style.icon, iconClassName)} type={currentIconType} />
-                    }
-                    {currentLabel}
-                  </React.Fragment>
-                }
-                {
-                  LinkComponent &&
-                  <LinkComponent {...{
-                    icon: currentIconType,
-                    label: currentLabel,
-                  }} />
-                }
-              </PageNavLink>
-            )
-          })}
-          {postAdditionalLinks.map(mapAdditionalLinks)}
-        </div>
-      )
-    }} />
-  )
-}
-
-HeaderMenu.propTypes = {
-  type: PropTypes.oneOf(Object.values(HEADER_TYPES)),
 }
 
 export { HEADER_TYPES }
