@@ -1,12 +1,16 @@
+import uuidV4 from 'uuid/v4'
+import debounce from 'lodash/debounce'
+
 import { api } from 'redux-restify'
 
+import auth from '$trood/auth'
 import { getToken } from '$trood/storage'
 import { objectToCamel, snakeToCamel } from '$trood/helpers/namingNotation'
 
 import {
   WS_URL,
   WS_PROTOCOL,
-  WS_OPEN_STATE,
+  WS_STATES,
 
   WS_MESSAGE_TYPES,
   WS_DOMAINS,
@@ -14,23 +18,26 @@ import {
 
   WS_PING_MESSAGE_TEXT,
   WS_PING_MESSAGE_INTERVAL,
+  WS_DEBOUNCE_INTERVAL,
 
-  WS_CLOSE_OK_CODE,
-  WS_CLOSE_OK_TEXT,
+  WS_CLOSE_ARGS,
 } from './constants'
 
 
-let socket = null
-let pingInterval = null
+let WS_ON_MESSAGE_CALLBACKS = {}
+let WS_SUBSCRIBES = {}
+let WS_MESSAGES = []
 
-const onMessage = (event, externalOnMessage) => dispatch => {
+let socket = null
+
+const onMessage = event => dispatch => {
   try {
     const dataParsed = objectToCamel(JSON.parse(event.data))
 
     if (Array.isArray(dataParsed) && dataParsed.length > 0) {
       dataParsed.forEach(item => {
-        if (item.data && item.domain === WS_DOMAINS.custodian) {
-          const { messageType, type, data } = item
+        const { domain, messageType, type, data } = item
+        if (data && data.id && domain === WS_DOMAINS.custodian) {
           const entityName = snakeToCamel(type)
 
           switch (messageType) {
@@ -45,7 +52,9 @@ const onMessage = (event, externalOnMessage) => dispatch => {
           }
         }
 
-        if (typeof externalOnMessage === 'function') externalOnMessage(item)
+        Object.values(WS_ON_MESSAGE_CALLBACKS).forEach(callback => {
+          if (typeof callback === 'function') callback(item)
+        })
       })
     }
   } catch (e) {
@@ -53,51 +62,72 @@ const onMessage = (event, externalOnMessage) => dispatch => {
   }
 }
 
-const sendData = data => {
-  if (socket) {
-    try {
-      const dataJSON = JSON.stringify(data)
-      socket.send(dataJSON)
-    } catch (e) {
-      console.warn('WS Data to JSON failed', e, data)
-    }
+const sendMessages = () => {
+  if (socket && socket.readyState === WS_STATES.open) {
+    WS_MESSAGES.forEach((data, i) => {
+      if (data) socket.send(data)
+      WS_MESSAGES[i] = null
+    })
+    WS_MESSAGES = WS_MESSAGES.filter(v => v)
   }
 }
+const debounceSendMessages = debounce(sendMessages, WS_DEBOUNCE_INTERVAL)
 
-export const init = (externalOnMessage, data) => dispatch => {
+const pushMessage = message => {
+  WS_MESSAGES.push(message)
+  debounceSendMessages()
+}
+
+const initAction = dispatch => {
   const token = getToken()
 
   if (token && !socket) {
-    socket = new WebSocket(`${WS_URL}?token=${token}`)
-    socket.onmessage = event => dispatch(onMessage(event, externalOnMessage))
-    socket.onopen = () => {
+    const ws = new WebSocket(`${WS_URL}?token=${token}`)
+    let pingInterval
+
+    ws.onopen = () => {
       pingInterval = setInterval(
-        () => sendData(WS_PING_MESSAGE_TEXT),
+        () => {
+          if (!WS_MESSAGES.length) pushMessage(WS_PING_MESSAGE_TEXT)
+        },
         WS_PING_MESSAGE_INTERVAL,
       )
-      sendData(data)
+      debounceSendMessages()
+      socket = ws
     }
-    socket.onclose = () => clearInterval(pingInterval)
-    socket.onerror = e => console.warn('An error in socket.onerror appeared', e)
+
+    ws.onclose = () => {
+      clearInterval(pingInterval)
+      socket = null
+    }
+
+    ws.onmessage = event => dispatch(onMessage(event))
+
+    ws.onerror = e => console.warn('An error in socket.onerror appeared', e)
+  }
+}
+const debounceInit = debounce(initAction, WS_DEBOUNCE_INTERVAL)
+const init = () => dispatch => debounceInit(dispatch)
+
+export const close = () => () => {
+  if (socket && socket.readyState === WS_STATES.open) {
+    socket.close(...WS_CLOSE_ARGS)
   }
 }
 
-export const close = () => () => {
-  if (socket && socket.readyState === WS_OPEN_STATE) socket.close(WS_CLOSE_OK_CODE, WS_CLOSE_OK_TEXT)
-}
+export const subscribe = ({
+  domain = WS_DOMAINS.custodian,
+  data = [],
+  onMessageCallback,
+} = {}) => (dispatch, getState) => {
+  dispatch(init())
 
-export const subscribe = (params = {}) => dispatch => {
-  const {
-    recipients = [],
-    domain = WS_DOMAINS.custodian,
-    data = [],
-
-    externalOnMessage,
-  } = params
-
+  const authData = auth.selectors.getAuthData(getState())
+  const subscribeUuid = uuidV4()
   const config = {
+    subscribeUuid,
     protocol: WS_PROTOCOL,
-    recipients: [...recipients],
+    recipients: [authData.login],
     data: {
       domain,
       action: WS_ACTIONS.subscribe,
@@ -105,5 +135,38 @@ export const subscribe = (params = {}) => dispatch => {
     },
   }
 
-  dispatch(init(externalOnMessage, config))
+  try {
+    const data = JSON.stringify(config)
+    pushMessage(data)
+    WS_SUBSCRIBES[subscribeUuid] = config
+    if (onMessageCallback) {
+      WS_ON_MESSAGE_CALLBACKS[subscribeUuid] = onMessageCallback
+    }
+    return subscribeUuid
+  } catch (e) {
+    console.warn('WS Data to JSON failed', e, config)
+  }
+  return undefined
+}
+
+export const unSubscribe = subscribeUuid => () => {
+  const subscribe = WS_SUBSCRIBES[subscribeUuid]
+  if (subscribe && socket && socket.readyState === WS_STATES.open) {
+    try {
+      const data = {
+        ...subscribe,
+        data: {
+          ...subscribe.data,
+          action: WS_ACTIONS.unsubscribe,
+        },
+      }
+      pushMessage(JSON.stringify(data))
+    } catch (e) {
+      console.warn('WS Data to JSON failed', e, subscribe)
+    }
+  } else {
+    WS_MESSAGES = WS_MESSAGES.filter(v => v.indexOf(subscribeUuid) < 0)
+  }
+  delete WS_SUBSCRIBES[subscribeUuid]
+  delete WS_ON_MESSAGE_CALLBACKS[subscribeUuid]
 }
